Remove jcenter from gradle; use parametrized test for crypto
