+++ build.gradle
@@ -3,9 +3,7 @@ buildscript {
         springBootVersion = '2.0.2.RELEASE'
     }
     repositories {
-        jcenter()
         mavenCentral()
-        maven { url "https://jitpack.io" }
     }
     dependencies {
         classpath("org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}")
