+++ src/test/java/jp/co/soramitsu/sora/crypto/algorithms/Ed25519Sha3SignatureTest.java
@@ -5,34 +5,66 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.security.KeyPair;
+import java.util.ArrayList;
+import java.util.List;
+import jp.co.soramitsu.crypto.ed25519.Utils;
 import jp.co.soramitsu.sora.crypto.algorithms.RawSignatureStrategy.SignatureSuiteException;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 
+@RunWith(Parameterized.class)
 public class Ed25519Sha3SignatureTest {
 
-  private RawSignatureStrategy sig;
+  private static final String fileName = "/crypto/test.data.sha3";
+  private RawSignatureStrategy sig = new Ed25519Sha3Signature();
+  private TestTuple testCase;
 
-  public Ed25519Sha3SignatureTest() throws SignatureSuiteException {
-    this.sig = new Ed25519Sha3Signature();
+  public Ed25519Sha3SignatureTest(TestTuple tuple) throws SignatureSuiteException {
+    this.testCase = tuple;
+  }
+
+  @Parameterized.Parameters
+  public static List<TestTuple> parameters() throws IOException {
+    List<TestTuple> testCases = new ArrayList<>();
+
+    InputStream is = Ed25519Sha3SignatureTest.class.getResourceAsStream(fileName);
+    if (is == null) {
+      throw new IOException("Resource not found: " + fileName);
+    }
+
+    BufferedReader file = new BufferedReader(new InputStreamReader(is));
+    String line;
+    while ((line = file.readLine()) != null) {
+      testCases.add(new TestTuple(line));
+    }
+    return testCases;
   }
 
   @Test
   public void rawSign() throws SignatureSuiteException {
-    for (TestVectors.TestTuple testCase : TestVectors.testCases) {
-      KeyPair keyPair = sig.generateKeypair(testCase.seed);
-      byte[] signature = sig.rawSign(testCase.message, keyPair);
+    KeyPair keyPair = sig.generateKeypair(testCase.seed);
+    byte[] signature = sig.rawSign(testCase.message, keyPair);
 
-      assertArrayEquals(testCase.sig, signature);
-    }
+    assertArrayEquals(testCase.sig, signature);
   }
 
   @Test
   public void rawVerify() throws SignatureSuiteException {
-    for (TestVectors.TestTuple testCase : TestVectors.testCases) {
-      KeyPair keyPair = sig.generateKeypair(testCase.seed);
-      assertTrue(sig.rawVerify(testCase.message, testCase.sig, keyPair.getPublic()));
-    }
+    KeyPair keyPair = sig.generateKeypair(testCase.seed);
+
+    assertTrue(
+        sig.rawVerify(
+            testCase.message,
+            testCase.sig,
+            keyPair.getPublic()
+        )
+    );
   }
 
   @Test
@@ -45,13 +77,11 @@ public class Ed25519Sha3SignatureTest {
 
   @Test
   public void generateKeypairFromSeed() throws SignatureSuiteException {
-    for (TestVectors.TestTuple testCase : TestVectors.testCases) {
-      KeyPair keyPair = sig.generateKeypair(testCase.seed);
+    KeyPair keyPair = sig.generateKeypair(testCase.seed);
 
-      assertNotNull(keyPair);
-      assertNotNull(keyPair.getPrivate());
-      assertNotNull(keyPair.getPublic());
-    }
+    assertNotNull(keyPair);
+    assertNotNull(keyPair.getPrivate());
+    assertNotNull(keyPair.getPublic());
   }
 
   @Test
@@ -60,5 +90,21 @@ public class Ed25519Sha3SignatureTest {
         "Ed25519Sha3Signature",
         sig.getType()
     );
+  }
+
+  public static class TestTuple {
+
+    public TestTuple(String line) {
+      String[] x = line.split(":");
+      seed = Utils.hexToBytes(x[0].substring(0, 64)); // private key
+      pk = Utils.hexToBytes(x[1]); // public key
+      message = Utils.hexToBytes(x[2]);
+      sig = Utils.hexToBytes(x[3].substring(0, 128)); // signature
+    }
+
+    private byte[] seed;
+    private byte[] pk;
+    private byte[] message;
+    private byte[] sig;
   }
 }
